+++ src/Container/HaystackStringLocate.php
@@ -17,15 +17,17 @@ class HaystackStringLocate
     /**
      * @param $value
      * @return int
+     * @throws ElementNotFoundException
+     * @throws \InvalidArgumentException
      */
     public function locate($value)
     {
-        if (is_scalar($value)) {
-            return $this->string->contains($value) ? strpos($this->string, (string) $value) : -1;
-        }
+        if (is_scalar($value) || $value instanceof HString) {
+            if ($this->string->contains($value)) {
+                return strpos($this->string, (string) $value);
+            }
 
-        if ($value instanceof HString) {
-            return $this->string->contains($value) ? strpos($this->string, $value->toString()) : -1;
+            throw new ElementNotFoundException($value);
         }
 
         throw new \InvalidArgumentException(sprintf("%s is neither a scalar value nor an HString", Helper::getType($value)));